scripts/pyproject.toml
[tool.poetry]
name = "data_preparation"
version = "0.1.0"
description = ""
authors = ["Sebastián Dalceggio <sdalceggio@gmail.com>"]
readme = "README.md"
packages = [{include = "data_preparation"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
xlrd = "^2.0.1"
pandas = "^2.0.2"
matplotlib = "^3.7.1"
python-decouple = "^3.8"
sqlalchemy = "1.4.44"
alembic = "^1.11.1"
pyodbc = "^4.0.39"
google-cloud-storage = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"
apache-airflow = "^2.6.1"
docker = "^6.1.3"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
